Remove navigation property from Player domain object
Remove the PlayerReference navigation property and replace it with a Guid that holds the id of the related player reference instead. This will cause the Name property to be turned into a method that loops over all available player references in tournament and find the related player reference by the id. Similar changes have been applied to other parts of the Player domain object